MANAGE: Fix unsigned comparison on update counts
This was causing the "bogus CN0 drop" bug, if the tracking channel
updated at the wrong time.

+++ src/manage.c
@@ -456,14 +456,14 @@ static void manage_track()
     u32 uc = ch->update_count;
 
     /* Optimistic phase lock detector "unlocked" for a while? */
-    if (uc - ch->ld_opti_locked_count > TRACK_DROP_UNLOCKED_T) {
+    if ((int)(uc - ch->ld_opti_locked_count) > TRACK_DROP_UNLOCKED_T) {
       log_info("PRN%d PLL unlocked too long, dropping\n", ch->prn+1);
       drop_channel(i);
       continue;
     }
 
     /* CN0 below threshold for a while? */
-    if (uc - ch->cn0_above_drop_thres_count > TRACK_DROP_CN0_T) {
+    if ((int)(uc - ch->cn0_above_drop_thres_count) > TRACK_DROP_CN0_T) {
       log_info("PRN%d low CN0 too long, dropping\n", ch->prn+1);
       drop_channel(i);
       continue;
@@ -482,7 +482,7 @@ s8 use_tracking_channel(u8 i)
       /* Some time has elapsed since the last tracking channel mode
        * change, to allow any transients to stabilize.
        * TODO: is this still necessary? */
-      && (ch->update_count - ch->mode_change_count > TRACK_STABILIZATION_T)
+      && ((int)(ch->update_count - ch->mode_change_count) > TRACK_STABILIZATION_T)
       /* Channel time of week has been decoded. */
       && (ch->TOW_ms != TOW_INVALID)
       /* Nav bit polarity is known, i.e. half-cycles have been resolved. */
